Deprecate KMSEnvelopeAEADKeyTemplate, introduce CreateKMSEnvelopeAEADKeyTemplate.
KMSEnvelopeAEADKeyTemplate is modified to panic if marshaling the DEK template proto fails. Previously this was silently ignored and would lead to subsequent failure when attempting to use the key template. This implementation is marked as deprecated.

CreateKSMEnvelopeAEADKeyTemplate is an equivalent function which returns an error value. If marshaling the DEK template proto fails, an error will be returned. This is the recommended replacement for the above, which avoids any potential for panic handling.

PiperOrigin-RevId: 486247274
Change-Id: I328bcc5aa0e50b68e8e62748a704cb3defc05aa2