Deprecate MemReaderWriter.
MemReaderWriter gives the user access to the keyset proto, which is usually not what they want. Instead, they want to convert the Handle from or to bytes or JSON strings, and they should Binary or JSON reader and writers instead.

Also, add tests that show how MemReaderWriter can be replaced.

PiperOrigin-RevId: 513464332
Change-Id: I526fcbb1667159ddf31fd15ea63d80043e2f4db3