Avoid re- or overallocating memory in HPKE helper function in golang.
If no capacity is specified, golang will choose a default capacity, usually 8 bytes. This means that if more than 8 bytes need to be appended, a new allocation needs to be done, and the content needs to be copied.

For these function, we already know excatly how many bytes we will need, so it is preferable to allocate the exact amount at the beginning.

PiperOrigin-RevId: 534158178
Change-Id: I078e39022f5e3f53b5774585d149e6fcb33baf14