Let keyset.NewBinaryWriter not write KeysetInfo to encrypted keysets in Golang.
KeysetInfo is unused, Tink never reads it and doesn't offer an API to access it. This change makes the Go implementation of Tink consistent with the C++ implementation where this is never written.

Note that the keyset.NewJsonWriter is unchanged and still does write KeysetInfo.

This change makes encrypted keyset considerably smaller: for example, the size of an encrypted keyset with a single AES 128 GCM key is reduced from 192 bytes to 122 bytes.

PiperOrigin-RevId: 596614620
Change-Id: I80c134c4cc5579d8742db6d215ceb091813f003c