Add AllPrimitives and AllPrimitivesWithKeyManager
These new methods on the keyset handle complement `Primitives` and
`PrimitivesWithKeyManager`, but return all primitives, not just enabled
primitives. This can be useful for key implementations to display
information about key parameters, but should not be used to create sets
for encryption duties, since only enabled keys should be used for
encryption.

Done by adding a filter function to the private `primitives` method, and
having `Primitives` and `PrimitivesWithKeyManager` use it to filter for
enabled keys, while `AllPrimitives` and `AllPrimitivesWithKeyManager`
don't filter. This keeps the code changes to a minimum. Also: add test
to ensure that disabled keys are included in the `All` methods but
excluded in the non-`All` methods.

Requires disabling the check for enabled keys in `primitiveset`; should
probably change it to allow the adding of other keys through a separate
method or something, but I wanted to get some feedback and discussion
going before I dove into that, so simply commenting out that check (and
the corresponding test) for now.