Remove embedded interface in streaming aead subtle implementations.
This embedding is not needed, because aesGCMHKDFSegmentDecrypter already implements this interface.

In fact, it is bad to add this embedding, because if there is for example a typo in the implementation of the interface, then the code still compiles but will panic at runtime because the implementation is nil. It is preferable that it doesn't compile.

PiperOrigin-RevId: 627065152
Change-Id: Iddd4b9e969294e70f91716ab69d8e1a108a8211e