Let aesGCMHKDFSegmentEncrypter.EncryptSegment use the correct result size.
Currently, the allocated memory for result has the same size of the plaintext, but the ciphertext is slightly larger and does not fit into result. Therefore, Seal will need to make another memory allocation. This change now increases the size of result so that the ciphertext fits.

Before this change, BenchmarkEncryptDecrypt/AES128_GCM_HKDF_4KB-8 allocates 56Mb/ops, after this change 34Mb/ops.

PiperOrigin-RevId: 627087400
Change-Id: Id5dcc5252ed92b2cb491b1744d551b93b758bd65