Make explicit copy when "append" is used to concatenate slices.
After this call:
a = append(b, c)
"a" may or may not use the same array as "b". So using "b" afterwards is error prone and should be avoided.

The usage of append in encrypt.go would be fine, because encapsulatedKey is not used anymore afterwards. But in this particular case, append anyways needs to make a copy, because the capacity of encapsulatedKey is not long enough. So it is better to make this copy explicit.

Once we drop support for go 1.21, we could replace this by slices.Concat that was added in Go 1.22. It would slightly simplify the code.

Also, fix a lint warning that removes an unnecessary nil check.

PiperOrigin-RevId: 631153171
Change-Id: I7b8291aa79a0c5e7075b70d2b7d56f259845ab43