Add nil-check for newWithOptions.
This makes sure that ks in a Handle is never nil. A handle with a nil ks doens't work properly anyways, and will result in an error or even a panic later anyways.

This check already exists for NewHandleWithNoSecrets, we also add a test for that.

This changes the behavior of the deprecated insecurecleartextkeyset.KeysetHandle function: calling insecurecleartextkeyset.KeysetHandle(nil) will now return nil instead of a keyset.Handle with ks=nil, which is basically unusable.

PiperOrigin-RevId: 635760959
Change-Id: I2be29d9ee0310754815d14bac0c32a9a5d7b9664