Better define ID requirement in keys
The ID requirement influences the behavior of the cryptographic function the key represents (e.g., it can be added to the output prefix). However, it only matches the keyset key ID if the ID is required (e.g., if the key has an output prefix).
Also update `aesgcm.Equals` to ignore the key ID requirements if an output prefix is not needed.

PiperOrigin-RevId: 668932992
Change-Id: Ic2cbd2f587026cde1a1d23ba8411fd327188a210