Add `protokeyserialization.KeySerialization` to represent a key serialization
This intends to replace `tinkpb.Keyset_Key` as the input/output of parser/serializer.

Using an intermediate representation has the advantage of having a more consistent representation of a `key.Key` as proto (that is, `tinkpb.KeyData` + ID requirement and output prefix), leaving callers to generate a `tinkpb.Keyset_Key` from this. Furthermore, it makes it easier to enforce `idRequirement == 0` when the key has no ID requirement for all keys, which can simply take this from granted.

PiperOrigin-RevId: 670498689
Change-Id: Iadecb3b9b251f4a1b162d1cccb2917efb898401f