Reintroduce `PrimitiveFromKeyData` in Config
Removing this API was premature; having a "legacy" way to create primitive using key manager vs the "new" way using constructor functions makes the logic to get `Primitives()` cleaner. The goal is to have `PrimitiveFromKey` always return "full" primitives, that is, primitives that are also aware of the output prefix, and `PrimitiveFromKeyData` to return legacy primitives that do not add any prefix to the output.

PiperOrigin-RevId: 689779144
Change-Id: I8cab76eafa1fb3ecac920f749c3c9c90c3a24fca