Use cipher.NewGCMWithRandomNonce in xaesgcm
For the same reasons as in aesgcm and with the same implementation, use
the new FIPS-assured GCM initialization in Go 1.24. This is a debatable
step because XAES-GCM is not a FIPS cryptographic primitive, but it is a
wrapper around the AES-GCM primitive and the Go Cryptographic module
ensures the latter is used in a FIPS-compliant way.

This change does not seek to answer the question of whether it should be
disabled, although reverting it would be a good way to disable it in the
GODEBUG=fips140=only mode of operation, as instantiation was already
failling with an error (albeit a rather opaque one that does not
indicate a FIPS compliance issue) in FIPS-only mode.