Add a new internal `KeyDeriver` interface and a PRF-based implementation
This is a primitive that creates a single `key.Key`. This primitive is "full" in that it generates a key with the a key ID if the key derivation key has an ID requirement.

NOTE: This currently doesn't support Streaming AEAD. This is currently unused by `wrappedKeysetDeriver`, so this is safe.
PiperOrigin-RevId: 773685451
Change-Id: Ib7ffb9fb15228b493e075102076f470096a3e0b5