Enable using full primitives in the JWT verifier factory
A "full" JWT verifier primitive implements `jwt.Verifier` and expects the correct kid.

The factory tries first to obtain a `jwt.Verifier` primitive from the keyset, and falls back to `verifierWithKID` if the first did not succeed.

NOTE: This assumes all the keys in the keyset will implement either `jwt.Verifier`, or `verifierWithKID` exclusively. Until all the keys are defined and the full primitives defined and registered, this is unused.
PiperOrigin-RevId: 794971534
Change-Id: Ie89f4ed96fba2e171b7cf3bdab6db4fb5aa69493