Replace "sha256sum -c" with "sha256sum -c -" wherever it might be called on MacOS.
Trying to run the tests on MacOS Sequoia failed with errors such as:

++ sha256sum -c
usage: sha256sum [-bctwz] [files ...]

According to pages such as https://issues.apache.org/jira/browse/MWRAPPER-150 we have that the version of sha256sum which ships with macOS Sequoia has a bug and requires an additional "-" to explicitly specify we read from stdin.

Hence we add that.

PiperOrigin-RevId: 797338234
Change-Id: I58fbe312b952cc1c94a289af5eccb0bdccd3aca9