Refactor `primitiveset.Entry` to use `key.Key`.
This removes the need to keep a proto key in the primitive set, which is "converted" on the fly primitive factories when needed (e.g., for signatures to know when the prefix is `LEGACY`).

It implicitly assumes that keys with an output prefix will implement an `OutputPrefix() []byte` method. This is true for all of the keys in Tink, but currently is not in the `key.Key` interface.

PiperOrigin-RevId: 816178408
Change-Id: Ie20e70c522144108ee3eeff33a60f3fbb080f387