Refactor fallback logic for legacy primitives
We remove `configWithLegacyFallback` and move the fallback logic to `registryconfig.RegistryConfig` to only allow using key managers when `registry` is used. We allow `keyset.Handle` to distinguish a legacy primitive from a full primitive using a new internal `legacyprimitive.LegacyPrimitive` type that `registryconfig` can use.

`LegacyPrimitive` is in a different `legacyprimitive` package than `registryconfig` to avoid a dependency of `keyset` on `registryconfig`.

PiperOrigin-RevId: 833278556
Change-Id: I4d8dd19edd5d1a2e110139a908cde7548f02d7e8